Minor package refactoring and increasing test coverage
