+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/PojoTest.java
@@ -9,8 +9,8 @@ import pl.pojo.tester.api.assertion.Method;
 import pl.pojo.tester.internal.field.DefaultFieldValueChanger;
 import pl.pojo.tester.internal.instantiator.ObjectGenerator;
 import pl.pojo.tester.internal.utils.ThoroughFieldPermutator;
-import uk.gov.hmcts.reform.wataskmanagementapi.auth.role.entities.GetRoleAssignmentResponse;
-import uk.gov.hmcts.reform.wataskmanagementapi.auth.role.entities.RoleAssignments;
+import uk.gov.hmcts.reform.wataskmanagementapi.auth.role.entities.request.RoleRequest;
+import uk.gov.hmcts.reform.wataskmanagementapi.auth.role.entities.response.GetRoleAssignmentResponse;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.AddLocalVariableRequest;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.HistoryVariableInstance;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.idam.Token;
@@ -32,7 +32,7 @@ class PojoTest {
         HistoryVariableInstance.class,
         AddLocalVariableRequest.class,
         GetRoleAssignmentResponse.class,
-        RoleAssignments.class,
+        RoleRequest.class
     };
 
     // Cannot test equals for generic classes
